backend/requirements/base.txt
# Spotify AI Agent – Base Requirements
# -----------------------------------
# Enthält alle Kernabhängigkeiten für Backend, ML, Security, DB, API, Microservices
# Rollen: Lead Dev, Architecte IA, Backend Senior, ML Engineer, DBA/Data Engineer, Security Specialist, Microservices Architect

fastapi>=0.110
uvicorn[standard]>=0.29
gunicorn>=21.2
python-dotenv>=1.0
pydantic>=2.7
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
redis>=5.0
celery>=5.4
httpx>=0.27
requests>=2.32
loguru>=0.7
structlog>=24.1
sentry-sdk>=2.0
prometheus-client>=0.20
opentelemetry-sdk>=1.25
opentelemetry-instrumentation>=0.45
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pyjwt>=2.8
bcrypt>=4.1
python-multipart>=0.0.9
PyYAML>=6.0
pymongo>=4.7
elasticsearch>=8.13
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.9
seaborn>=0.13
# ML/AI
transformers>=4.41
torch>=2.3
tensorflow>=2.16
# Spotify API
spotipy>=2.23
# Compliance
python-json-logger>=2.0
